+++ tensorzero-internal/tests/e2e/best_of_n.rs
@@ -6,9 +6,10 @@ use tensorzero_internal::inference::{
 };
 use uuid::Uuid;
 
-use crate::common::{
-    get_clickhouse, get_gateway_endpoint, select_chat_inference_clickhouse,
-    select_json_inference_clickhouse, select_model_inferences_clickhouse,
+use crate::common::get_gateway_endpoint;
+use tensorzero_internal::clickhouse::test_helpers::{
+    get_clickhouse, select_chat_inference_clickhouse, select_json_inference_clickhouse,
+    select_model_inferences_clickhouse,
 };
 
 /// This test calls a function which currently uses best of n.
