+++ tensorzero-internal/tests/e2e/inference.rs
@@ -21,11 +21,13 @@ use tensorzero_internal::{
 };
 use uuid::Uuid;
 
+use tensorzero_internal::clickhouse::test_helpers::{
+    get_clickhouse, select_chat_inference_clickhouse, select_json_inference_clickhouse,
+    select_model_inference_clickhouse,
+};
+
 use crate::{
-    common::{
-        get_clickhouse, get_gateway_endpoint, select_chat_inference_clickhouse,
-        select_json_inference_clickhouse, select_model_inference_clickhouse,
-    },
+    common::get_gateway_endpoint,
     providers::common::{
         make_embedded_gateway, make_embedded_gateway_no_config, make_http_gateway,
     },
