+++ tensorzero-internal/tests/e2e/mixture_of_n.rs
@@ -3,9 +3,10 @@ use serde_json::{json, Value};
 use tensorzero_internal::inference::types::{ContentBlock, RequestMessage, Role};
 use uuid::Uuid;
 
-use crate::common::{
-    get_clickhouse, get_gateway_endpoint, select_chat_inference_clickhouse,
-    select_json_inference_clickhouse, select_model_inferences_clickhouse,
+use crate::common::get_gateway_endpoint;
+use tensorzero_internal::clickhouse::test_helpers::{
+    get_clickhouse, select_chat_inference_clickhouse, select_json_inference_clickhouse,
+    select_model_inferences_clickhouse,
 };
 
 /// This test calls a function which currently uses mixture of n.
