+++ tensorzero-internal/tests/e2e/providers/anthropic.rs
@@ -5,11 +5,11 @@ use serde_json::{json, Value};
 use uuid::Uuid;
 
 use crate::{
-    common::{
-        get_clickhouse, get_gateway_endpoint, select_chat_inference_clickhouse,
-        select_model_inference_clickhouse,
-    },
+    common::get_gateway_endpoint,
     providers::common::{E2ETestProvider, E2ETestProviders},
+};
+use tensorzero_internal::clickhouse::test_helpers::{
+    get_clickhouse, select_chat_inference_clickhouse, select_model_inference_clickhouse,
 };
 
 #[cfg(feature = "e2e_tests")]
