+++ tensorzero-internal/tests/e2e/providers/batch.rs
@@ -11,7 +11,9 @@ use reqwest::{Client, StatusCode};
 use serde_json::{json, Value};
 use std::collections::HashSet;
 use tensorzero_internal::{
-    clickhouse::ClickHouseConnectionInfo,
+    clickhouse::{
+        test_helpers::select_batch_model_inferences_clickhouse, ClickHouseConnectionInfo,
+    },
     endpoints::batch_inference::PollPathParams,
     inference::types::{
         batch::{BatchModelInferenceRow, BatchRequestRow},
@@ -23,25 +25,23 @@ use tokio::time::{sleep, Duration};
 use url::Url;
 use uuid::Uuid;
 
-use crate::{
-    common::select_batch_model_inferences_clickhouse,
-    providers::common::{
-        check_dynamic_json_mode_inference_response, check_dynamic_tool_use_inference_response,
-        check_json_mode_inference_response, check_parallel_tool_use_inference_response,
-        check_tool_use_multi_turn_inference_response,
-        check_tool_use_tool_choice_allowed_tools_inference_response,
-        check_tool_use_tool_choice_auto_unused_inference_response,
-        check_tool_use_tool_choice_auto_used_inference_response,
-        check_tool_use_tool_choice_none_inference_response,
-        check_tool_use_tool_choice_required_inference_response,
-        check_tool_use_tool_choice_specific_inference_response,
-    },
+use tensorzero_internal::clickhouse::test_helpers::{
+    get_clickhouse, select_batch_model_inference_clickhouse, select_latest_batch_request_clickhouse,
+};
+
+use crate::providers::common::{
+    check_dynamic_json_mode_inference_response, check_dynamic_tool_use_inference_response,
+    check_json_mode_inference_response, check_parallel_tool_use_inference_response,
+    check_tool_use_multi_turn_inference_response,
+    check_tool_use_tool_choice_allowed_tools_inference_response,
+    check_tool_use_tool_choice_auto_unused_inference_response,
+    check_tool_use_tool_choice_auto_used_inference_response,
+    check_tool_use_tool_choice_none_inference_response,
+    check_tool_use_tool_choice_required_inference_response,
+    check_tool_use_tool_choice_specific_inference_response,
 };
 use crate::{
-    common::{
-        get_clickhouse, get_gateway_endpoint, select_batch_model_inference_clickhouse,
-        select_latest_batch_request_clickhouse,
-    },
+    common::get_gateway_endpoint,
     providers::common::{check_inference_params_response, check_simple_inference_response},
 };
