+++ tensorzero-internal/tests/e2e/retries.rs
@@ -11,9 +11,10 @@ use tensorzero_internal::inference::{
 };
 use uuid::Uuid;
 
-use crate::common::{
-    get_clickhouse, get_gateway_endpoint, select_chat_inference_clickhouse,
-    select_model_inference_clickhouse, select_model_inferences_clickhouse,
+use crate::common::get_gateway_endpoint;
+use tensorzero_internal::clickhouse::test_helpers::{
+    get_clickhouse, select_chat_inference_clickhouse, select_model_inference_clickhouse,
+    select_model_inferences_clickhouse,
 };
 
 /// This test calls a function which calls a model where the provider is flaky but with retries.
